deps: modularize the data-generator dependency
